Moved convert functions to helper file
